requirements.txt
pydantic
streamlit
google-genai
python-dotenv
